+++ commands/container-registry/repository/get.go
@@ -1,0 +1,87 @@
+package repository
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/ionos-cloud/ionosctl/v6/commands/container-registry/registry"
+	"github.com/ionos-cloud/ionosctl/v6/internal/client"
+	"github.com/ionos-cloud/ionosctl/v6/internal/constants"
+	"github.com/ionos-cloud/ionosctl/v6/internal/core"
+	"github.com/ionos-cloud/ionosctl/v6/internal/printer/json2table/jsonpaths"
+	"github.com/ionos-cloud/ionosctl/v6/internal/printer/jsontabwriter"
+	"github.com/ionos-cloud/ionosctl/v6/internal/printer/tabheaders"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+func RepositoryGetCmd() *core.Command {
+	c := core.NewCommand(
+		context.Background(), nil, core.CommandBuilder{
+			Namespace:  "container-registry",
+			Resource:   "repository",
+			Verb:       "get",
+			ShortDesc:  "Retrieve a repository.",
+			LongDesc:   "Retrieve a specific repository from a registry.",
+			Example:    "ionosctl container-registry get",
+			PreCmdRun:  PreCmdGet,
+			CmdRun:     CmdGet,
+			InitClient: true,
+		},
+	)
+	c.Command.Flags().StringSlice(constants.ArgCols, defaultCols, tabheaders.ColsMessage(allCols))
+	_ = c.Command.RegisterFlagCompletionFunc(
+		constants.ArgCols,
+		func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return allCols, cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+
+	c.AddStringFlag(constants.FlagRegistryId, constants.FlagRegistryIdShort, "", "Registry ID")
+	_ = c.Command.RegisterFlagCompletionFunc(
+		constants.FlagRegistryId,
+		func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return registry.RegsIds(), cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+
+	c.AddStringFlag(constants.FlagName, constants.FlagNameShort, "", "Name of the repository to get")
+	_ = c.Command.RegisterFlagCompletionFunc(
+		constants.FlagName,
+		func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			return RepositoryNames(
+				viper.GetString(core.GetFlagName(c.NS, constants.FlagRegistryId)),
+			), cobra.ShellCompDirectiveNoFileComp
+		},
+	)
+
+	return c
+}
+
+func PreCmdGet(c *core.PreCommandConfig) error {
+	return core.CheckRequiredFlags(c.Command, c.NS, constants.FlagRegistryId, constants.FlagName)
+}
+
+func CmdGet(c *core.CommandConfig) error {
+	cols, _ := c.Command.Command.Flags().GetStringSlice(constants.ArgCols)
+	regId := viper.GetString(core.GetFlagName(c.NS, constants.FlagRegistryId))
+	name := viper.GetString(core.GetFlagName(c.NS, constants.FlagName))
+
+	repo, _, err := client.Must().RegistryClient.RepositoriesApi.RegistriesRepositoriesFindByName(
+		context.
+			Background(), regId, name,
+	).Execute()
+	if err != nil {
+		return err
+	}
+
+	out, err := jsontabwriter.GenerateOutput(
+		"", jsonpaths.ContainerRegistryRepository, repo, tabheaders.GetHeaders(allCols, defaultCols, cols),
+	)
+	if err != nil {
+		return err
+	}
+
+	fmt.Fprintf(c.Command.Command.OutOrStdout(), out)
+	return nil
+}