+++ commands/container-registry/token/constants.go
@@ -1,10 +1,8 @@
 package token
 
 const (
-	FlagRegId           = "registry-id"
 	FlagTokenId         = "token-id"
 	FlagAllTokens       = "all-tokens"
-	FlagName            = "name"
 	FlagExpiryDate      = "expiry-date"
 	FlagTimeUntilExpiry = "expiry-time"
 	FlagStatus          = "status"